requirements.txt
Django==1.10
djangorestframework==3.4.5
numpy
matplotlib
scipy
sklearn
hmmlearn
simplejson
eyed3
pygame
pyaudio
